Cleaned up some of the comments and removed some warning messages so that now everything compiles clean with VC++ warning level 3

+++ mpeg1decoder/vlc.h
@@ -82,8 +82,8 @@ namespace Mpeg1
 		static const short s_macroblockTypeP[];
 		static const short s_macroblockTypeB[];
 
-		static const short s_motionVector[];
-		static const short s_motionVector1[];
+		static const int s_motionVector[];
+		static const int s_motionVector1[];
 
 		static const short s_codedBlockPattern[];
